+++ server/user/src/main/kotlin/com/ring/ring/user/withdrawal/WithdrawalUser.kt
@@ -1,22 +1,25 @@
 package com.ring.ring.user.withdrawal
 
 import com.ring.ring.user.UseCase
-import com.ring.ring.user.ValidateSession
+import com.ring.ring.user.shared.NotLoggedInException
+import com.ring.ring.user.shared.Session
+import com.ring.ring.user.shared.ValidateSessionRepository
 import kotlinx.serialization.Serializable
 
 class WithdrawalUser(
-    private val validateSession: ValidateSession = ValidateSession(),
+    private val sessionRepository: ValidateSessionRepository = ValidateSessionRepository(),
     private val repository: WithdrawalUserRepository = WithdrawalUserModules.withdrawalUserRepository,
 ) : UseCase<WithdrawalUser.Req, WithdrawalUser.Res>() {
     override suspend fun execute(req: Req): Res {
-        validateSession(req.session)
+        val isValid = sessionRepository.validate(req.session)
+        if (isValid.not()) throw NotLoggedInException()
         repository.delete(id = req.session.userId)
         return Res()
     }
 
     @Serializable
     data class Req(
-        val session: ValidateSession.ReqSession,
+        val session: Session,
     ) : UseCase.Req
 
     class Res : UseCase.Res
